Create PullUpManFinalProject.py
The final project for our camp connect campers, to create an interactive ascii-art based console game in python. 
The game is similar to hangman in that the player will guess letters of a secret word, but instead of getting closer to drawing a hangman, 
in this game the player gets closer and closer to helping pull-up man do a pull-up!